Remove partial_by_order
