+++ dask/tests/test_utils.py
@@ -1,5 +1,6 @@
-import pickle
 import functools
+import operator
+import pickle
 
 import numpy as np
 import pytest
@@ -8,7 +9,7 @@ from dask.sharedict import ShareDict
 from dask.utils import (takes_multiple_arguments, Dispatch, random_state_data,
                         memory_repr, methodcaller, M, skip_doctest,
                         SerializableLock, funcname, ndeepmap, ensure_dict,
-                        extra_titles, asciitable, itemgetter)
+                        extra_titles, asciitable, itemgetter, partial_by_order)
 from dask.utils_test import inc
 
 
@@ -311,3 +312,7 @@ def test_itemgetter():
     g2 = pickle.loads(pickle.dumps(g))
     assert g2(data) == 2
     assert g2.index == 1
+
+
+def test_partial_by_order():
+    assert partial_by_order(5, function=operator.add, other=[(1, 20)]) == 25
