+++ dask/utils.py
@@ -899,3 +899,18 @@ def infer_storage_options(urlpath, inherit_storage_options=None):
         inferred_storage_options.update(inherit_storage_options)
 
     return inferred_storage_options
+
+
+def partial_by_order(*args, **kwargs):
+    """
+
+    >>> from operator import add
+    >>> partial_by_order(5, function=add, other=[(1, 10)])
+    15
+    """
+    function = kwargs.pop('function')
+    other = kwargs.pop('other')
+    args2 = list(args)
+    for i, arg in other:
+        args2.insert(i, arg)
+    return function(*args2, **kwargs)
